chore: remove dead functional helpers batch 002
