+++ docs/internal/baselines/deadcode-baseline.txt
@@ -83,20 +83,6 @@ pkg/services/work/transports/cli/work/list.go: unreachable func: List
 pkg/services/work/transports/cli/work/move.go: unreachable func: Move
 pkg/services/work/transports/cli/work/show.go: unreachable func: Show
 pkg/services/work/transports/cli/work/watch.go: unreachable func: RenderWatchTransition
-tests/functional/bootstrap_portability/automat_portability_fixture_test.go: unreachable func: assertAutomatBundledFileEntryWithoutInline
-tests/functional/bootstrap_portability/export_import_fixture_test.go: unreachable func: exportImportFixture.persistAs
-tests/functional/cli/mcp_resume/serve_runtime_resume_smoke_test.go: unreachable func: mcpControlResume
-tests/functional/cli/mcp_resume/stdio_client_test.go: unreachable func: assertInstallSmokeDiscovery
-tests/functional/internal/support/api_server.go: unreachable func: FunctionalAPIServer.WaitForExitError
-tests/functional/internal/support/api_server.go: unreachable func: StartFunctionalAPIServiceModeServer
-tests/functional/internal/support/events.go: unreachable func: LastFactoryEventTick
-tests/functional/internal/support/process.go: unreachable func: CreateAndActivateNamedFactoryWithProcess
-tests/functional/internal/support/process.go: unreachable func: createNamedFactoryAtRoot
-tests/functional/internal/support/process_api_server.go: unreachable func: ProcessAPIServer.Ready
-tests/functional/internal/support/process_factory.go: unreachable func: RunFactoryToCompletionWithEdgesAndObservationsStable
-tests/functional/internal/support/provider.go: unreachable func: CursorProviderSuccessStdout
-tests/functional/internal/support/provider.go: unreachable func: ProviderErrorCommandResult
-tests/functional/internal/support/workdir.go: unreachable func: WithWorkingDirectory
 tests/functional/orchestration/javascript/durability/resume_test.go: unreachable func: resumeJavaScriptSessionExpectingInvalidState
 tests/functional/orchestration/javascript/events/events_test.go: unreachable func: decodeNDJSONRecords
 tests/functional/orchestration/javascript/events/events_test.go: unreachable func: mapKeys
