+++ tests/functional/internal/support/workdir.go
@@ -33,26 +33,3 @@ func SetWorkingDirectory(t *testing.T, dir string) {
 		workingDirectoryMu.Unlock()
 	})
 }
-
-func WithWorkingDirectory(t *testing.T, dir string, fn func()) {
-	t.Helper()
-
-	workingDirectoryMu.Lock()
-	originalDir, err := os.Getwd()
-	if err != nil {
-		workingDirectoryMu.Unlock()
-		t.Fatalf("Getwd(): %v", err)
-	}
-	if err := os.Chdir(dir); err != nil {
-		workingDirectoryMu.Unlock()
-		t.Fatalf("Chdir(%q): %v", dir, err)
-	}
-	defer func() {
-		if err := os.Chdir(originalDir); err != nil {
-			t.Fatalf("restore working directory: %v", err)
-		}
-		workingDirectoryMu.Unlock()
-	}()
-
-	fn()
-}
